lagde owner objektet

+++ src/main/java/com/example/jimbobspethospital/Pets/petRepo.java
@@ -2,5 +2,5 @@ package com.example.jimbobspethospital.Pets;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface petRepo extends JpaRepository<pet, Integer> {
+public interface petRepo extends JpaRepository<pet, Long> {
 }
